refactor: create global context for watchlist
