+++ app/src/MyStocks/index.js
@@ -1,12 +1,14 @@
 import * as React from "react";
 
 import useApi from "../auth/useApi";
+import { useMyWatchlist } from "../myWatchlist";
 
 import Portfolios from "./Portfolios";
 import Watchlist from "./Watchlist";
 import styles from "./styles.module.scss";
 
-const MyStocks = ({ watchlist, setWatchlist }) => {
+const MyStocks = () => {
+  const { watchlist, setWatchlist } = useMyWatchlist();
   const { loading } = useApi();
 
   return loading ? null : (
