+++ app/src/MyStocks/styles.module.scss
@@ -36,7 +36,7 @@
     .portfolioCard {
       margin-left: auto;
       margin-right: auto;
-      max-width: 70%;
+      max-width: 60%;
       display: flex;
       flex-direction: column;
       justify-content: center;
