+++ app/src/index.js
@@ -5,6 +5,7 @@ import { BrowserRouter as Router } from "react-router-dom";
 
 import App from "./App";
 import AuthProvider from "./auth/AuthProvider";
+import { MyWatchlistProvider } from "./myWatchlist";
 import reportWebVitals from "./reportWebVitals";
 
 import "./global.scss";
@@ -13,7 +14,9 @@ ReactDOM.render(
   <React.StrictMode>
     <Router>
       <AuthProvider>
-        <App />
+        <MyWatchlistProvider>
+          <App />
+        </MyWatchlistProvider>
       </AuthProvider>
     </Router>
   </React.StrictMode>,
